requirements.txt
Flask==2.3.2
pdfplumber==0.10.4
pdf2image==1.16.3
pikepdf==9.5.2
flask-cors==4.0.0
werkzeug==2.3.7
PyPDF2==3.0.1
fitz==0.0.1.dev2
frontend
tools
pymupdf
docx2pdf
python-docx
pdf2docx
comtypes
pywin32
Pillow
python-pptx
openpyxl
pdfkit